requirements.txt
pandas
ortools
google-generativeai
scikit-learn
reportlab
numpy
